74 - more tidy up
