--- src/app/Core/services/table/table.service.ts
@@ -1,7 +1,0 @@
-import { Injectable } from '@angular/core'
-import PocketBase from 'pocketbase'
-
-@Injectable()
-export class TableService {
-	constructor() {}
-}
