+++ src/app/app.module.ts
@@ -22,6 +22,7 @@ import { TableComponent } from './Core/components/table/table.component'
 import { ApiService } from './Core/services/api/api.service'
 import { AuthGuardService } from './Core/services/auth/auth-guard.service'
 import { ErrorService } from './Core/services/error/error.service'
+import { FormService } from './Core/services/form/form.service'
 import { LoginService } from './Core/services/login/login.service'
 import { NotificationService } from './Core/services/notification/notification.service'
 import { QueryService } from './Core/services/query/query.service'
@@ -83,6 +84,7 @@ import { UsersComponent } from './Users/users/users.component'
 		SocialService,
 		QueryService,
 		ErrorService,
+		FormService,
 	],
 	bootstrap: [AppComponent],
 })
